Create or load existing file arrays

Create or load existing file arrays

filearray_create(
  filebase,
  dimension,
  type = c("double", "float", "integer", "logical", "raw", "complex"),
  partition_size = NA
)

filearray_load(filebase, mode = c("readwrite", "readonly"))

Arguments

filebase
a directory path to store arrays in the local file system. When creating an array, the path must not exist.

dimension
dimension of the array, at least length of 2

type
storage type of the array; default is 'double'. Other options include 'integer', 'logical', and 'raw'.

partition_size
positive partition size for the last margin, or NA to automatically guess; see 'Details'.

mode
whether allows writing to the file; choices are 'readwrite' and 'readonly'.

Value

A FileArray-class instance.

Details

The file arrays partition out-of-memory array objects and store them separately in local file systems. Since R stores matrices/arrays in column-major style, file array uses the slowest margin (the last margin) to slice the partitions. This helps to align the elements within the files with the corresponding memory order. An array with dimension 100x200x300x400 has 4 margins. The length of the last margin is 400, which is also the maximum number of potential partitions. The number of partitions are determined by the last margin size divided by partition_size. For example, if the partition size is 1, then there will be 400 partitions. If the partition size if 3, there will be 134 partitions. The default partition sizes are determined internally following these priorities:

1.

the file size of each partition does not exceed 1GB

2.

the number of partitions do not exceed 100

These two rules are not hard requirements. The goal is to reduce the numbers of partitions as much as possible.

Author

Zhengjia Wang

Examples

# Prepare 
library(filearray)
filebase <- tempfile()
if(file.exists(filebase)){ unlink(filebase, TRUE) }

# create array
x <- filearray_create(filebase, dimension = c(200, 30, 8))
print(x)

#> Reference class object of class "FileArray"
#> Mode: readwrite 
#> Dimension: 200x30x8 
#> # of partitions: 8 
#> Partition size: 1 
#> Storage type: double (internal size: 8)
#> Location: /tmp/RtmpHYRebz/file2d391d8b70fa 

# Assign values
x[] <- rnorm(48000)

# Subset
x[1,2,]

#> [1] -0.42938009  0.44742147  1.05363668 -0.97346749 -0.09327726 -1.14067120
#> [7]  1.45111737  0.47259544

# load existing array
filearray_load(filebase)

#> Reference class object of class "FileArray"
#> Mode: readwrite 
#> Dimension: 200x30x8 
#> # of partitions: 8 
#> Partition size: 1 
#> Storage type: double (internal size: 8)
#> Location: /tmp/RtmpHYRebz/file2d391d8b70fa