Set or get file array threads

Will enable/disable multi-threaded reading or writing at C++ level using 'OpenMP'.

filearray_threads(n, reset_after_fork = -1L)

Arguments

n

number of threads to set. If n is negative, then default to the number of cores that computer has.

reset_after_fork

when forking a cluster (in 'osx' or 'linux'), the number of threads will be set to one to avoid memory issues. Setting this to 1 will reset threads once forked clusters are shut down. Setting to 0 will disable reset, and -1 to use last set values.

Value

An integer of current number of threads